+++ csd2c/theorie/session_01/02_tremeloProcessBuffer/oscillators/saw.cpp
@@ -6,7 +6,7 @@
 Saw::Saw() : Saw(0, 0)
 {}
 
-Saw::Saw(double frequency, double samplerate) :
+Saw::Saw(float frequency, float samplerate) :
   Oscillator(frequency, samplerate),
   phaseIncrement(frequency / samplerate * SAW_SMOOTH_FACTOR),
   smoothY(0)
